+++ src/file_tree.mli
@@ -2,6 +2,24 @@
 
 open! Import
 
+module Dune_file : sig
+  module Plain : sig
+    (** [sexps] is mutable as we get rid of the S-expressions once
+        they have been parsed, in order to release the memory as soon
+        as we don't need them. *)
+    type t =
+      { path          : Path.t
+      ; mutable sexps : Sexp.Ast.t list
+      }
+  end
+
+  type t =
+    | Plain of Plain.t
+    | Ocaml_script of Path.t
+
+  val path : t -> Path.t
+end
+
 module Dir : sig
   type t
 
@@ -12,8 +30,8 @@ module Dir : sig
   val sub_dir_paths : t -> Path.Set.t
   val sub_dir_names : t -> String.Set.t
 
-  (** Whether this directory is ignored by a [jbuild-ignore] file in
-      one of its ancestor directories. *)
+  (** Whether this directory is ignored by an [ignored_subdirs] stanza
+     or [jbuild-ignore] file in one of its ancestor directories. *)
   val ignored : t -> bool
 
   val fold
@@ -23,8 +41,8 @@ module Dir : sig
     -> f:(t -> 'a -> 'a)
     -> 'a
 
-  (** Return the dune (or jbuild) file in this directory *)
-  val dune_file : t -> Path.t option
+  (** Return the contents of the dune (or jbuild) file in this directory *)
+  val dune_file : t -> Dune_file.t option
 end
 
 (** A [t] value represent a view of the source tree. It is lazily
@@ -37,7 +55,7 @@ val load : ?extra_ignored_subtrees:Path.Set.t -> Path.t -> t
 
 (** Passing [~traverse_ignored_dirs:true] to this functions causes the
     whole source tree to be deeply scanned, including ignored
-    directories. *)
+    sub-trees. *)
 val fold
   :  t
   -> traverse_ignored_dirs:bool