+++ CHANGES.md
@@ -40,6 +40,9 @@ next
 - Fix a crash when using c files from another directory (#758, fixes
   #734, @diml)
 
+- Add an `ignored_subdirs` stanza to replace `jbuild-ignore` files
+  (#767, @diml)
+
 1.0+beta20 (10/04/2018)
 -----------------------